% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GN1000.R
\name{GN1000}
\alias{GN1000}
\title{Sample groundnut PGR passport data}
\format{A data frame having 1000 records with the following 10 
  columns(fields): \itemize{ \item{CommonName}{ : }{Common name} 
  \item{BotanicalName}{ : }{Botanical name} \item{NationalID}{ : }{NBPGR 
  National identifier} \item{CollNo}{ : }{Collector number} \item{DonorID}{ :
  }{Donor ID} \item{OtherID1}{ : }{Other ID field 1} \item{OtherID2}{ : 
  }{Other ID field 2} \item{BioStatus}{ : }{Biological status}
  \item{SourceCountry}{ : }{Country of origin} \item{TransferYear}{ : }{Year
  of transfer} }}
\usage{
data(GN1000)
}
\description{
Sample PGR passport data of 1000 groundnut accessions held in the Indian 
National Genebank at National Bureau of Plant Genetic Resources (NBPGR), New 
Delhi.
}
\seealso{
\url{http://www.nbpgr.ernet.in:8080/PGRPortal/}
}
\keyword{datasets}
