% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.KWIC.R
\name{print.KWIC}
\alias{print.KWIC}
\title{Prints summary of \code{KWIC} object.}
\usage{
\method{print}{KWIC}(x, ...)
}
\arguments{
\item{x}{An object of class \code{KWIC}.}

\item{...}{Unused}
}
\description{
\code{print.KWIC} prints to console the summary of an object of class 
\code{KWIC} including the database fields(columns) used, the total number of
keywords and the number of distinct keywords in the index.
}
\seealso{
\code{\link[PGRdup]{KWIC}}
}
