% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ProbDup.R
\name{print.ProbDup}
\alias{print.ProbDup}
\title{Prints summary of \code{ProbDup} object.}
\usage{
\method{print}{ProbDup}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ProbDup}.}

\item{...}{Unused}
}
\description{
\code{print.ProbDup} prints to console the summary of an object of class 
\code{ProbDup} including the method used ("a", "b" or "c"), the database
fields(columns) considered, the number of probable duplicate sets of each
kind along with the corresponding number of records.
}
\seealso{
\code{\link[PGRdup]{ProbDup}}
}
