% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGRdup-package.R
\docType{package}
\name{PGRdup-package}
\alias{PGRdup-package}
\alias{PGRdup}
\title{The PGRdup Package}
\description{
Functions to facilitate genebank managers in the identification of probable 
duplicate accessions from plant genetic resources (PGR) passport databases.
}
\author{
J Aravind \email{aravindj@nbpgr.ernet.in} \cr J Radhamani 
  \email{radhamani@nbpgr.ernet.in} \cr Kalyani Srinivasan 
  \email{kalyani@nbpgr.ernet.in} \cr B Ananda Subhash 
  \email{anandasubhash@gmail.com} \cr RK Tyagi
  \email{rktyagi@nbpgr.ernet.in}
}
